json data handle adjustments
